refactor: split out parsing logic into module
